% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_subgroup_effects.R
\name{combine_summaries}
\alias{combine_summaries}
\title{Combine Multiple Subgroup Summaries for Comparison Plotting}
\usage{
combine_summaries(summary_list)
}
\arguments{
\item{summary_list}{Named list of `subgroup_summary` objects.}
}
\value{
`subgroup_summary`. Combined object with a "Model" column in estimates.
}
\description{
Combines estimates from multiple subgroup summary objects into a single
summary object that can be plotted to compare models.
}
\examples{
\dontrun{
summary1 <- summary_subgroup_effects(brms_fit = model1)
summary2 <- summary_subgroup_effects(brms_fit = model2)

combined <- combine_summaries(list(
  "OVAT" = summary1,
  "Global" = summary2
))
plot(combined)
}

}
