% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_subgroup_effects.R
\name{plot.subgroup_summary}
\alias{plot.subgroup_summary}
\title{Plot Marginal Subgroup Treatment Effects}
\usage{
\method{plot}{subgroup_summary}(x, x_lab = NULL, title = NULL, ...)
}
\arguments{
\item{x}{An object of class `subgroup_summary`.}

\item{x_lab}{A character string for the x-axis label.}

\item{title}{A character string for the plot title.}

\item{...}{Additional arguments.}
}
\value{
A `ggplot` object representing the forest plot.
}
\description{
Creates a forest plot from a `subgroup_summary` object.
}
