% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsaiforest2-package.R
\docType{package}
\name{bonsaiforest2-package}
\alias{bonsaiforest2}
\alias{bonsaiforest2-package}
\title{bonsaiforest2: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{internal}
