% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_subgroup_effects.R
\name{plot.subgroup_summary}
\alias{plot.subgroup_summary}
\title{Plot Marginal Subgroup Treatment Effects}
\usage{
\method{plot}{subgroup_summary}(x, x_lab = NULL, title = NULL, ...)
}
\arguments{
\item{x}{`subgroup_summary`. Object created by `summary_subgroup_effects()`.}

\item{x_lab}{`character(1)` or `NULL`. Custom label for x-axis.}

\item{title}{`character(1)` or `NULL`. Custom title for plot.}

\item{...}{Additional arguments (currently unused).}
}
\value{
`ggplot`. Forest plot visualization of subgroup treatment effects.
}
\description{
Creates a forest plot from a `subgroup_summary` object.
}
