% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_subgroup_effects.R
\name{summary_subgroup_effects}
\alias{summary_subgroup_effects}
\title{Create a Summary of Marginal Subgroup Treatment Effects}
\usage{
summary_subgroup_effects(
  brms_fit,
  original_data,
  trt_var,
  response_type = c("binary", "count", "continuous", "survival"),
  subgroup_vars = "auto"
)
}
\arguments{
\item{brms_fit}{A fitted `brmsfit` object.}

\item{original_data}{The original `data.frame` used for model fitting.}

\item{trt_var}{The name of the treatment variable (coded 0/1).}

\item{response_type}{The type of outcome variable. One of "binary", "count",
"continuous", or "survival".}

\item{subgroup_vars}{A character vector of subgrouping variable names.
Defaults to "auto". Cannot be NULL.}
}
\value{
An object of class `subgroup_summary`.
}
\description{
This function orchestrates calls to `estimate_subgroup_effects` to generate
a summary table for specific subgroups.
}
